# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Roleva API",
    "description": "Evidence-based resume analysis.",
    "version": "0.1.0"
  },
  "paths": {
    "/analyze": {
      "post": {
        "summary": "Analyze",
        "description": "Analyse one resume against one job description.\n\nSynchronous. At fewer than a hundred analyses a day there is no queue worth\nrunning, and a request that returns the finished report is far simpler to\nreason about than a job id the client has to poll.",
        "operationId": "analyze_analyze_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_analyze_analyze_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisReport": {
        "additionalProperties": false,
        "properties": {
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "resume": {
            "$ref": "#/components/schemas/ResumeDocument"
          },
          "job": {
            "$ref": "#/components/schemas/JobTarget"
          },
          "matches": {
            "$ref": "#/components/schemas/MatchReport"
          },
          "ats": {
            "$ref": "#/components/schemas/AtsReport"
          },
          "scores": {
            "$ref": "#/components/schemas/ScoreReport"
          },
          "headline": {
            "default": "",
            "title": "Headline",
            "type": "string"
          },
          "verdict": {
            "default": "",
            "title": "Verdict",
            "type": "string"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "title": "Strengths",
            "type": "array"
          },
          "weaknesses": {
            "items": {
              "type": "string"
            },
            "title": "Weaknesses",
            "type": "array"
          },
          "section_feedback": {
            "items": {
              "$ref": "#/components/schemas/SectionFeedback"
            },
            "title": "Section Feedback",
            "type": "array"
          },
          "bullet_suggestions": {
            "items": {
              "$ref": "#/components/schemas/BulletSuggestion"
            },
            "title": "Bullet Suggestions",
            "type": "array"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "rubric_version": {
            "title": "Rubric Version",
            "type": "string"
          },
          "prompt_version": {
            "title": "Prompt Version",
            "type": "string"
          },
          "degraded_stages": {
            "items": {
              "$ref": "#/components/schemas/Stage"
            },
            "title": "Degraded Stages",
            "type": "array"
          }
        },
        "required": [
          "id",
          "status",
          "created_at",
          "resume",
          "job",
          "matches",
          "ats",
          "scores",
          "rubric_version",
          "prompt_version"
        ],
        "title": "AnalysisReport",
        "type": "object"
      },
      "AnalysisStatus": {
        "enum": [
          "pending",
          "running",
          "complete",
          "partial",
          "failed"
        ],
        "title": "AnalysisStatus",
        "type": "string"
      },
...
      "AtsReport": {
        "additionalProperties": false,
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/AtsFinding"
            },
            "title": "Findings",
            "type": "array"
          },
          "hidden_text_detected": {
            "default": false,
            "title": "Hidden Text Detected",
            "type": "boolean"
          },
          "checks_run": {
            "default": 0,
            "title": "Checks Run",
            "type": "integer"
          }
        },
        "title": "AtsReport",
        "type": "object"
      },
...
      "Body_analyze_analyze_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "The resume, as a PDF"
          },
          "job_description": {
            "type": "string",
            "title": "Job Description",
            "description": "The full job posting text"
          }
        },
        "type": "object",
        "required": [
          "file",
          "job_description"
        ],
        "title": "Body_analyze_analyze_post"
      },
...
      "Bullet": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Span"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "required": [
          "text"
        ],
        "title": "Bullet",
        "type": "object"
      },
      "BulletSuggestion": {
        "additionalProperties": false,
        "description": "A concrete improvement for one weak bullet.\n\nGrounding rule: the suggestion may not introduce facts (numbers, employers,\ntechnologies) absent from `original`. Violations are regenerated once, then\ndropped.",
        "properties": {
          "bullet_id": {
            "title": "Bullet Id",
            "type": "string"
          },
          "original": {
            "title": "Original",
            "type": "string"
          },
          "suggestion": {
            "title": "Suggestion",
            "type": "string"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "title": "Reasons",
            "type": "array"
          },
          "section": {
            "$ref": "#/components/schemas/SectionKind"
          }
        },
        "required": [
          "bullet_id",
          "original",
          "suggestion",
          "section"
        ],
        "title": "BulletSuggestion",
        "type": "object"
      },
      "ContactInfo": {
        "additionalProperties": false,
        "description": "Populated by deterministic extraction. Feeds the PII redactor \u2014 every\nvalue here is masked before any text is sent to an external LLM.",
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Phone"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Location"
          },
          "links": {
            "items": {
              "type": "string"
            },
            "title": "Links",
            "type": "array"
          }
        },
        "title": "ContactInfo",
        "type": "object"
      },
      "CredentialItem": {
        "additionalProperties": false,
        "description": "Certifications, awards, publications \u2014 same shape, different section.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "issuer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Issuer"
          },
          "dates": {
            "$ref": "#/components/schemas/DateRange"
          },
          "span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Span"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "required": [
          "title"
        ],
        "title": "CredentialItem",
        "type": "object"
      },
...
      "DetectedSection": {
        "additionalProperties": false,
        "description": "A heading found in the source document, with its text range.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/SectionKind"
          },
          "heading_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Heading Text"
          },
          "order": {
            "default": 0,
            "title": "Order",
            "type": "integer"
          },
          "span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Span"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "heading_confidence": {
            "default": 1.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Heading Confidence",
            "type": "number"
          }
        },
        "required": [
          "kind"
        ],
        "title": "DetectedSection",
        "type": "object"
      },
      "EducationItem": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "degree": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Degree"
          },
          "field_of_study": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Field Of Study"
          },
          "institution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Institution"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Location"
          },
          "dates": {
            "$ref": "#/components/schemas/DateRange"
          },
          "gpa": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Gpa"
          },
          "bullets": {
            "items": {
              "$ref": "#/components/schemas/Bullet"
            },
            "title": "Bullets",
            "type": "array"
          },
          "span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Span"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "title": "EducationItem",
        "type": "object"
      },
...
      "ExpectedBand": {
        "additionalProperties": false,
        "description": "Reference range for a role family + seniority. Curated, not measured \u2014\nalways labelled as a \"typical range\", never as a percentile.",
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ScoreKind"
          },
          "low": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Low",
            "type": "number"
          },
          "high": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "High",
            "type": "number"
          },
          "position": {
            "title": "Position",
            "type": "string"
          }
        },
        "required": [
          "kind",
          "low",
          "high",
          "position"
        ],
        "title": "ExpectedBand",
        "type": "object"
      },
      "ExperienceItem": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Title"
          },
          "organization": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Organization"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Location"
          },
          "dates": {
            "$ref": "#/components/schemas/DateRange"
          },
          "bullets": {
            "items": {
              "$ref": "#/components/schemas/Bullet"
            },
            "title": "Bullets",
            "type": "array"
          },
          "span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Span"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "title": "ExperienceItem",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobTarget": {
        "additionalProperties": false,
        "description": "A parsed job description.",
        "properties": {
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Title"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Company"
          },
          "role_family": {
            "default": "general",
            "title": "Role Family",
            "type": "string"
          },
          "seniority": {
            "$ref": "#/components/schemas/Seniority",
            "default": "unknown"
          },
          "requirements": {
            "items": {
              "$ref": "#/components/schemas/Requirement"
            },
            "title": "Requirements",
            "type": "array"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "title": "JobTarget",
        "type": "object"
      },
      "MatchReport": {
        "additionalProperties": false,
        "properties": {
          "matches": {
            "items": {
              "$ref": "#/components/schemas/RequirementMatch"
            },
            "title": "Matches",
            "type": "array"
          },
          "unmatched_resume_skills": {
            "items": {
              "type": "string"
            },
            "title": "Unmatched Resume Skills",
            "type": "array"
          },
          "llm_adjudicated_count": {
            "default": 0,
            "title": "Llm Adjudicated Count",
            "type": "integer"
          }
        },
        "title": "MatchReport",
        "type": "object"
      },
...
      "ProjectItem": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Description"
          },
          "technologies": {
            "items": {
              "type": "string"
            },
            "title": "Technologies",
            "type": "array"
          },
          "dates": {
            "$ref": "#/components/schemas/DateRange"
          },
          "bullets": {
            "items": {
              "$ref": "#/components/schemas/Bullet"
            },
            "title": "Bullets",
            "type": "array"
          },
          "link": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Link"
          },
          "span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Span"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "title": "ProjectItem",
        "type": "object"
      },
...
      "Recommendation": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "projected_gain": {
            "default": 0.0,
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Projected Gain",
            "type": "number"
          },
          "related_requirement_ids": {
            "items": {
              "type": "string"
            },
            "title": "Related Requirement Ids",
            "type": "array"
          }
        },
        "required": [
          "id",
          "title",
          "detail",
          "severity"
        ],
        "title": "Recommendation",
        "type": "object"
      },
      "Requirement": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "canonical": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Canonical"
          },
          "category": {
            "$ref": "#/components/schemas/RequirementCategory"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "priority_source": {
            "$ref": "#/components/schemas/Provenance",
            "default": "extracted"
          },
          "quantifier": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Quantifier"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "mention_count": {
            "default": 1,
            "minimum": 1,
            "title": "Mention Count",
            "type": "integer"
          },
          "span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Span"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "required": [
          "text",
          "category",
          "priority"
        ],
        "title": "Requirement",
        "type": "object"
      },
...
      "RequirementMatch": {
        "additionalProperties": false,
        "description": "The resolved outcome for a single JD requirement.",
        "properties": {
          "requirement_id": {
            "title": "Requirement Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MatchStatus"
          },
          "strength": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Strength",
            "type": "number"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "title": "Evidence",
            "type": "array"
          },
          "is_adjacent": {
            "default": false,
            "title": "Is Adjacent",
            "type": "boolean"
          },
          "observed_months": {
            "anyOf": [
              {
                "minimum": 0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Observed Months"
          },
          "explanation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Explanation"
          }
        },
        "required": [
          "requirement_id",
          "status",
          "strength"
        ],
        "title": "RequirementMatch",
        "type": "object"
      },
      "ResumeDocument": {
        "additionalProperties": false,
        "description": "The structured resume. Analyzer reads it; the future builder will edit it.",
        "properties": {
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "contact": {
            "$ref": "#/components/schemas/ContactInfo"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Bullet"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "experience": {
            "items": {
              "$ref": "#/components/schemas/ExperienceItem"
            },
            "title": "Experience",
            "type": "array"
          },
          "education": {
            "items": {
              "$ref": "#/components/schemas/EducationItem"
            },
            "title": "Education",
            "type": "array"
          },
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectItem"
            },
            "title": "Projects",
            "type": "array"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/SkillMention"
            },
            "title": "Skills",
            "type": "array"
          },
          "certifications": {
            "items": {
              "$ref": "#/components/schemas/CredentialItem"
            },
            "title": "Certifications",
            "type": "array"
          },
          "awards": {
            "items": {
              "$ref": "#/components/schemas/CredentialItem"
            },
            "title": "Awards",
            "type": "array"
          },
          "publications": {
            "items": {
              "$ref": "#/components/schemas/CredentialItem"
            },
            "title": "Publications",
            "type": "array"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/DetectedSection"
            },
            "title": "Sections",
            "type": "array"
          },
          "page_count": {
            "default": 0,
            "minimum": 0,
            "title": "Page Count",
            "type": "integer"
          },
          "parse_confidence": {
            "default": 0.0,
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Parse Confidence",
            "type": "number"
          },
          "parse_warnings": {
            "items": {
              "type": "string"
            },
            "title": "Parse Warnings",
            "type": "array"
          }
        },
        "title": "ResumeDocument",
        "type": "object"
      },
...
      "ScoreReport": {
        "additionalProperties": false,
        "properties": {
          "overall": {
            "$ref": "#/components/schemas/Score"
          },
          "job_match": {
            "$ref": "#/components/schemas/Score"
          },
          "ats": {
            "$ref": "#/components/schemas/Score"
          },
          "quality": {
            "$ref": "#/components/schemas/Score"
          },
          "must_coverage": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Must Coverage",
            "type": "number"
          },
          "overall_coverage": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Overall Coverage",
            "type": "number"
          },
          "expected_bands": {
            "items": {
              "$ref": "#/components/schemas/ExpectedBand"
            },
            "title": "Expected Bands",
            "type": "array"
          },
          "percentiles": {
            "items": {
              "$ref": "#/components/schemas/Percentile"
            },
            "title": "Percentiles",
            "type": "array"
          },
          "rubric_version": {
            "title": "Rubric Version",
            "type": "string"
          }
        },
        "required": [
          "overall",
          "job_match",
          "ats",
          "quality",
          "must_coverage",
          "overall_coverage",
          "rubric_version"
        ],
        "title": "ScoreReport",
        "type": "object"
      },
      "SectionFeedback": {
        "additionalProperties": false,
        "properties": {
          "section": {
            "$ref": "#/components/schemas/SectionKind"
          },
          "rating": {
            "maximum": 5,
            "minimum": 1,
            "title": "Rating",
            "type": "integer"
          },
          "strengths": {
            "items": {
              "type": "string"
            },
            "title": "Strengths",
            "type": "array"
          },
          "issues": {
            "items": {
              "type": "string"
            },
            "title": "Issues",
            "type": "array"
          }
        },
        "required": [
          "section",
          "rating"
        ],
        "title": "SectionFeedback",
        "type": "object"
      },
...
      "Seniority": {
        "enum": [
          "intern",
          "entry",
          "mid",
          "senior",
          "unknown"
        ],
        "title": "Seniority",
        "type": "string"
      },
...
      "SkillMention": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "raw": {
            "title": "Raw",
            "type": "string"
          },
          "canonical": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Canonical"
          },
          "origin": {
            "$ref": "#/components/schemas/SkillOrigin",
            "default": "other"
          },
          "span": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Span"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          }
        },
        "required": [
          "raw"
        ],
        "title": "SkillMention",
        "type": "object"
      },
...
      "Stage": {
        "description": "Pipeline stages. Names are user-visible in the progress UI, so they\ndescribe work the user cares about, not internal function names.",
        "enum": [
          "validating",
          "extracting",
          "structuring",
          "reading_job",
          "matching",
          "checking_ats",
          "assessing_quality",
          "scoring",
          "writing_advice",
          "done"
        ],
        "title": "Stage",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}